refactor(security-utils): separate log safety ownership
