+++ scripts/background/handlers/saveHandlers.js
@@ -14,8 +14,8 @@ import {
   validateContentScriptRequest,
   isValidNotionUrl,
   sanitizeApiError,
-  sanitizeUrlForLogging,
 } from '../../utils/securityUtils.js';
+import { sanitizeUrlForLogging } from '../../utils/LogSanitizer.js';
 import { buildHighlightBlocks } from '../utils/BlockBuilder.js';
 import { parseArgsToContext } from '../../utils/Logger.js';
 import {
